clean

+++ src/board/utils.rs
@@ -1,4 +1,6 @@
-pub fn factorial(num: u64) -> u64 {
+use crate::algo::graph::{apply_action, new_position, Dir};
+
+pub fn factorial(num: u128) -> u128 {
 	match num {
 		0 | 1 => 1,
 		_ => match factorial(num - 1).checked_mul(num) {
